Features

+++ app/src/main/java/com/example/meditake/database/dao/ProgrammeDao.java
@@ -21,7 +21,7 @@ public interface ProgrammeDao {
     long insert(Programme p);
 
     @Query("Select * from Programme where id = :id")
-    Programme getById(int id);
+    Programme getById(long id);
 
     @Query("SELECT * FROM Programme WHERE id IN (:userIds)")
     List<Programme> loadAllByIds(int[] userIds);
