+++ app/src/main/java/com/example/meditake/database/dao/RapportDao.java
@@ -21,7 +21,7 @@ public interface RapportDao {
     long insert(Rapport p);
 
     @Query("Select * from Rapport where id = :id")
-    Rapport getById(int id);
+    Rapport getById(long id);
 
     @Query("SELECT * FROM Rapport WHERE id IN (:userIds)")
     List<Rapport> loadAllByIds(int[] userIds);
